backEnd/framework/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djongo==1.3.6
Markdown==3.3.4
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
